refactor: App 내부를 모듈화
